+++ src/internal-links/handler.js
@@ -11,12 +11,15 @@
  */
 
 import RUMAPIClient from '@adobe/spacecat-shared-rum-api-client';
+import { internalServerError } from '@adobe/spacecat-shared-http-utils';
 import { getRUMDomainkey, getRUMUrl } from '../support/utils.js';
 import { AuditBuilder } from '../common/audit-builder.js';
 import { noopUrlResolver } from '../common/audit.js';
+import { syncSuggestions } from '../utils/data-access.js';
 
 const INTERVAL = 30; // days
 const DAILY_PAGEVIEW_THRESHOLD = 100;
+const AUDIT_TYPE = 'broken-internal-links';
 
 /**
  * Determines if the URL has the same host as the current host.
@@ -107,7 +110,77 @@ export async function internalLinksAuditRunner(auditUrl, context, site) {
   };
 }
 
+// eslint-disable-next-line consistent-return
+export async function convertToOpportunity(auditUrl, auditData, context) {
+  const {
+    dataAccess,
+    log,
+  } = context;
+
+  let opportunity;
+  try {
+    const opportunities = await dataAccess.Opportunity.allBySiteIdAndStatus(auditData.siteId, 'NEW');
+    opportunity = opportunities.find((oppty) => oppty.getType() === AUDIT_TYPE);
+  } catch (e) {
+    log.error(`Fetching opportunities for siteId ${auditData.siteId} failed with error: ${e.message}`);
+    return internalServerError(`Failed to fetch opportunities for siteId ${auditData.siteId}: ${e.message}`);
+  }
+
+  try {
+    if (!opportunity) {
+      const opportunityData = {
+        siteId: auditData.siteId,
+        auditId: auditData.id,
+        runbook: 'https://adobe.sharepoint.com/sites/aemsites-engineering/Shared%20Documents/3%20-%20Experience%20Success/SpaceCat/Runbooks/Experience_Success_Studio_Broken_Internal_Links_Runbook.docx?web=1',
+        type: AUDIT_TYPE,
+        origin: 'AUTOMATION',
+        title: 'Broken internal links found',
+        description: 'We\'ve detected broken internal links on your website. Broken links can negatively impact user experience and SEO. Please review and fix these links to ensure smooth navigation and accessibility.',
+        guidance: {
+          steps: [
+            'Update each broken internal link to valid URLs.',
+            'Test the implemented changes manually to ensure they are working as expected.',
+            'Monitor internal links for 404 errors in RUM tool over time to ensure they are functioning correctly.',
+          ],
+        },
+        tags: [
+          'Traffic acquisition',
+          'Engagement',
+        ],
+      };
+      opportunity = await dataAccess.Opportunity.create(opportunityData);
+    } else {
+      opportunity.setAuditId(auditData.id);
+      await opportunity.save();
+    }
+  } catch (e) {
+    log.error(`Failed to create new opportunity for siteId ${auditData.siteId} and auditId ${auditData.id}: ${e.message}`);
+    throw e;
+  }
+
+  const buildKey = (item) => `${item.url_from}-${item.url_to}`;
+
+  // Sync suggestions
+  await syncSuggestions({
+    opportunity,
+    newData: auditData?.auditResult?.brokenInternalLinks,
+    buildKey,
+    mapNewSuggestion: (entry) => ({
+      opportunityId: opportunity.getId(),
+      type: 'CONTENT_UPDATE',
+      rank: entry.traffic_domain,
+      data: {
+        ...entry,
+        /* code commented until implementation of suggested links. TODO: implement suggestions, https://jira.corp.adobe.com/browse/SITES-26545 */
+        // suggestedLink: 'some suggestion here',
+      },
+    }),
+    log,
+  });
+}
+
 export default new AuditBuilder()
   .withUrlResolver(noopUrlResolver)
   .withRunner(internalLinksAuditRunner)
+  .withPostProcessors([convertToOpportunity])
   .build();