+++ test/audits/cwv.test.js
@@ -15,12 +15,15 @@
 import { expect, use } from 'chai';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
+import chaiAsPromised from 'chai-as-promised';
 import nock from 'nock';
-import { createSite } from '@adobe/spacecat-shared-data-access/src/models/site.js';
-import { CWVRunner } from '../../src/cwv/handler.js';
-import { rumData } from '../fixtures/rum-data.js';
+import { CWVRunner, convertToOppty } from '../../src/cwv/handler.js';
+import expectedOppty from '../fixtures/cwv/oppty.json' with { type: 'json' };
+import suggestions from '../fixtures/cwv/suggestions.json' with { type: 'json' };
+import rumData from '../fixtures/cwv/cwv.json' with { type: 'json' };
 
 use(sinonChai);
+use(chaiAsPromised);
 
 const sandbox = sinon.createSandbox();
 
@@ -28,19 +31,29 @@ const baseURL = 'https://spacecat.com';
 const auditUrl = 'www.spacecat.com';
 const DOMAIN_REQUEST_DEFAULT_PARAMS = {
   domain: auditUrl,
-  domainkey: 42,
+  domainkey: '42',
   interval: 7,
   granularity: 'hourly',
 };
+const AUDIT_TYPE = 'cwv';
 
-describe('Index Tests', () => {
-  const site = createSite({ baseURL });
+describe('CWVRunner Tests', () => {
+  const groupedURLs = [{ test: 'test' }];
+  const siteConfig = {
+    getGroupedURLs: sandbox.stub().returns(groupedURLs),
+  };
+  const site = {
+    getBaseURL: sandbox.stub().returns(baseURL),
+    getConfig: () => siteConfig,
+  };
+
   const context = {
     runtime: { name: 'aws-lambda', region: 'us-east-1' },
     func: { package: 'spacecat-services', version: 'ci', name: 'test' },
     rumApiClient: {
-      query: sandbox.stub().withArgs('variable-1', sinon.match(DOMAIN_REQUEST_DEFAULT_PARAMS)).resolves(rumData),
+      query: sandbox.stub().resolves(rumData),
     },
+    dataAccess: {},
   };
 
   beforeEach('setup', () => {
@@ -59,7 +72,19 @@ describe('Index Tests', () => {
   });
 
   it('cwv audit runs rum api client cwv query', async () => {
-    const result = await CWVRunner('www.spacecat.com', context, site);
+    const result = await CWVRunner(auditUrl, context, site);
+
+    expect(siteConfig.getGroupedURLs.calledWith(AUDIT_TYPE)).to.be.true;
+    expect(
+      context.rumApiClient.query.calledWith(
+        AUDIT_TYPE,
+        {
+          ...DOMAIN_REQUEST_DEFAULT_PARAMS,
+          groupedURLs,
+        },
+      ),
+    ).to.be.true;
+
     expect(result).to.deep.equal({
       auditResult: {
         cwv: rumData.filter((data) => data.pageviews >= 7000),
@@ -68,6 +93,121 @@ describe('Index Tests', () => {
         },
       },
       fullAuditRef: auditUrl,
+    });
+  });
+
+  describe('CWV audit to oppty conversion', () => {
+    let addSuggestionsResponse;
+    let oppty;
+    const opptyData = { 0: 'existed-data' };
+    let auditData;
+
+    beforeEach(() => {
+      context.log = {
+        info: sandbox.stub(),
+        error: sandbox.stub(),
+      };
+
+      context.dataAccess.Opportunity = {
+        allBySiteIdAndStatus: sandbox.stub(),
+        create: sandbox.stub(),
+      };
+
+      addSuggestionsResponse = {
+        createdItems: [],
+        errorItems: [],
+      };
+
+      oppty = {
+        getType: () => AUDIT_TYPE,
+        getId: () => 'oppty-id',
+        getSiteId: () => 'site-id',
+        addSuggestions: sandbox.stub().resolves(addSuggestionsResponse),
+        getSuggestions: sandbox.stub().resolves([]),
+        setAuditId: sandbox.stub(),
+        getData: sandbox.stub().returns(opptyData),
+        setData: sandbox.stub(),
+        save: sandbox.stub().resolves(),
+      };
+
+      auditData = {
+        siteId: 'site-id',
+        id: 'audit-id',
+        isLive: true,
+        auditedAt: new Date().toISOString(),
+        auditType: AUDIT_TYPE,
+        auditResult: {
+          cwv: rumData.filter((data) => data.pageviews >= 7000),
+          auditContext: {
+            interval: 7,
+          },
+        },
+        fullAuditRef: auditUrl,
+      };
+    });
+
+    afterEach(() => {
+      sandbox.restore();
+    });
+
+    it('creates a new opportunity object', async () => {
+      context.dataAccess.Opportunity.allBySiteIdAndStatus.resolves([]);
+      context.dataAccess.Opportunity.create.resolves(oppty);
+
+      await convertToOppty(auditUrl, auditData, context);
+
+      expect(context.dataAccess.Opportunity.create).to.have.been.calledOnceWith(expectedOppty);
+
+      // make sure that newly oppty has all 4 new suggestions
+      expect(oppty.addSuggestions).to.have.been.calledOnce;
+      const suggestionsArg = oppty.addSuggestions.getCall(0).args[0];
+      expect(suggestionsArg).to.be.an('array').with.lengthOf(4);
+    });
+
+    it('creating a new opportunity object fails', async () => {
+      context.dataAccess.Opportunity.allBySiteIdAndStatus.resolves([]);
+      context.dataAccess.Opportunity.create.rejects(new Error('big error happened'));
+
+      await expect(convertToOppty(auditUrl, auditData, context)).to.be.rejectedWith('big error happened');
+
+      expect(context.dataAccess.Opportunity.create).to.have.been.calledOnceWith(expectedOppty);
+      expect(context.log.error).to.have.been.calledOnceWith('Failed to create new opportunity for siteId site-id and auditId audit-id: big error happened');
+
+      // make sure that no new suggestions are added
+      expect(oppty.addSuggestions).to.have.been.to.not.have.been.called;
+    });
+
+    it('updates the existing opportunity object', async () => {
+      context.dataAccess.Opportunity.allBySiteIdAndStatus.resolves([oppty]);
+      const existingSuggestions = suggestions.map((suggestion) => ({
+        ...suggestion,
+        opportunityId: oppty.getId(),
+        remove: sinon.stub(),
+        save: sinon.stub(),
+        getData: () => (suggestion.data),
+        setData: sinon.stub(),
+      }));
+      oppty.getSuggestions.resolves(existingSuggestions);
+
+      await convertToOppty(auditUrl, auditData, context);
+
+      expect(context.dataAccess.Opportunity.create).to.not.have.been.called;
+      expect(oppty.setAuditId).to.have.been.calledOnceWith('audit-id');
+      expect(oppty.setData).to.have.been.calledOnceWith({ ...opptyData, ...expectedOppty.data });
+      expect(oppty.save).to.have.been.calledOnce;
+
+      // make sure that 1 old suggestion is removed
+      expect(existingSuggestions[0].remove).to.have.been.calledOnce;
+
+      // make sure that 1 existing suggestion is updated
+      expect(existingSuggestions[1].setData).to.have.been.calledOnce;
+      expect(existingSuggestions[1].setData.firstCall.args[0]).to.deep.equal(suggestions[1].data);
+      expect(existingSuggestions[1].save).to.have.been.calledOnce;
+
+      // make sure that 3 new suggestions are created
+      expect(oppty.addSuggestions).to.have.been.calledOnce;
+      const suggestionsArg = oppty.addSuggestions.getCall(0).args[0];
+      expect(suggestionsArg).to.be.an('array').with.lengthOf(3);
     });
   });
 });